requirements.txt
# pymirai==0.4.0
pillow==7.2.0
lxml==4.5.2
aiohttp==3.6.2
cairocffi==1.0.1
weasyprint
pydantic==1.6.1
graia-application-mirai

# pytest
# pytest-asyncio
# pytest-cov
# coverage<5

# mypy
